Move useTheme module to a folder, add notes type
