+++ src/lib/types/note.ts
@@ -1,0 +1,4 @@
+export type Note = {
+    title: string;
+    description: string;
+};
